Changed the priority of the requirement identification process to be lower than JacksonAnnotationIntrospector
closes https://github.com/FasterXML/jackson-module-kotlin/issues/941
